Make body types correct.
This was added as definitions aren't required by the spec to have a type,
but they are required to have a title. So a validating spec wouldn't
get definition flowtypes generated.

It turns out that hack is too general, as parameters can have titles too.

Need to figure out a different way to whitelist definitions.